size of little spheres, line width
On Mac the line width of lines is getting rendered correctly, on
Windows not (!). So had to set it back to 1 to look good on Mac.
The size of the little spheres (where the lines do start and end)
correlates now to the factor of inputC length / resultC length. Other
correlations will be explored…